+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.0.3 - Not Released
-- Stuff done
+- Reset the pagination when searching
 
 ## 1.0.2 - 2021-07-06
 ### Changed